requirements.txt
Flask
Flask-SocketIO
gspread
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
gevent
gevent-websocket
gunicorn
python-dotenv
